requirements.txt
requests>=2.32.3
python-docx>=1.1.2
jsonschema>=4.23.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.0.0
rich>=13.0.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
